requirements.txt
# Core CLI
textual>=0.89.0
click==8.1.7
rich==13.7.0
typer==0.9.0
pyfiglet==1.0.2

# AI/ML
mistralai==1.2.5
elevenlabs==1.11.0

# Data
datasets==3.2.0
pandas==2.2.3

# Monitoring
wandb>=0.22.3
weave==0.52.28

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0

# MCP
mcp>=1.0.0

# Audio playback (voice_narrator skill)
sounddevice>=0.5.0
numpy>=1.26.0

# Auth / Cloud
supabase>=2.0.0

# Utilities
pydantic==2.10.6
python-dotenv==1.0.1
pyyaml==6.0.2
requests==2.32.3
aiohttp==3.11.11
